move measurements function to do_insert_data
